Fixed all bugs
