+++ src/index.html
@@ -10,7 +10,7 @@
 <body>
   <div class="root">
     <header class="header root__section">
-      <img src="./images/logo.svg" alt="mesto logo" class="logo">
+      <img src="<%=require('./images/logo.svg')%>" alt="mesto logo" class="logo">
     </header>
     <div class="profile root__section">
       <div class="user-info">
@@ -61,7 +61,7 @@
 
     <div class="popup popup_place">
       <div class="popup__content">
-        <img src="./images/close.svg" alt="" class="popup__close popup__close_place">
+        <img src="<%=require('./images/close.svg')%>" alt="" class="popup__close popup__close_place">
         <h3 class="popup__title">Новое место</h3>
         <form class="popup__form" name="new">
           <input type="text" name="name" class="popup__input popup__input_type_name" placeholder="Название">
@@ -73,11 +73,11 @@
 
     <div id="popup_image" class="popup popup_image">
       <div class="popup__container">
-        <img src="./images/close.svg" alt="" class="popup__close popup__close_image">
+        <img src="<%=require('./images/close.svg')%>" alt="" class="popup__close popup__close_image">
       </div>
     </div>
     
   </div>
-  <script src="<%= htmlWebpackPlugin.files.chunks.main.entry %>"></script><!--Если сломается исправить на <script src="./main.js"></script>-->
+  <script type="module" src="<%= htmlWebpackPlugin.files.chunks.main.entry %>"></script><!--Если сломается исправить на <script src="./main.js"></script>-->
 </body>
 </html>
