+++ src/var.js
@@ -1,0 +1,5 @@
+const userInfoName = document.querySelector('.user-info__name');
+const userInfoJob = document.querySelector('.user-info__job');// работа в профиле пользователя
+const placesList = document.querySelector('.places-list');//контейнер для попапов с картинками
+const array = [];
+export {userInfoName, userInfoJob, placesList, array};
