+++ webpack.config.js
@@ -1,8 +1,10 @@
-const path = require('path');//Путь
-const MiniCssExtractPlugin = require("mini-css-extract-plugin"); // Подключили к проекту плагин
+const path = require('path');
 const HtmlWebpackPlugin = require('html-webpack-plugin');
 const WebpackMd5Hash = require('webpack-md5-hash');
+const MiniCssExtractPlugin = require('mini-css-extract-plugin');
 const webpack = require('webpack');
+const OptimizeCssAssetsPlugin = require('optimize-css-assets-webpack-plugin');
+const isDev = process.env.NODE_ENV === 'development';
 
 module.exports = {
     entry: { main: './src/index.js' },
@@ -17,28 +19,53 @@ module.exports = {
             exclude: /node_modules/ // исключает папку node_modules НУЖНА
             },
             {
-            test: /\.css$/, // применять это правило только к CSS-файлам
-            use: [MiniCssExtractPlugin.loader, 'css-loader', 'postcss-loader'] // к этим файлам нужно применить пакеты, которые мы уже установили 
+            test: /\.css$/i,
+            use: [
+                (isDev ? 'style-loader' : MiniCssExtractPlugin.loader),
+                'css-loader', 
+                'postcss-loader'
+                ]
             },
             {
             test: /\.(png|jpg|gif|ico|svg)$/,
-            use: {
-                loader: 'file-loader?name=./images/[name].[ext]'
-            },
+            use: [
+                'file-loader?name=../images/[name].[ext]', // указали папку, куда складывать изображения
+                {
+                    loader: 'image-webpack-loader',
+                    options: {}
+                },
+            ],
             },
             {
-                test: /\.(eot|ttf|woff|woff2)$/,
+            test: /\.(eot|ttf|woff|woff2)$/,
+            use: {
                 loader: 'file-loader?name=./fonts/[name].[ext]',
+            },
             }
             
         ]
     },
     plugins: [ 
-        new MiniCssExtractPlugin({filename: './css/[name].[contenthash].css'}),
-        new HtmlWebpackPlugin({ // настроили плагин
+        new MiniCssExtractPlugin({
+            filename: 'index.[contenthash].css',
+          }),
+        new HtmlWebpackPlugin({
             inject: false,
+            hash: true,
             template: './src/index.html',
-            filename: 'index.html'
-        })
+            filename: 'index.html',
+          }),
+        new WebpackMd5Hash(),
+        new webpack.DefinePlugin({
+            NODE_ENV: JSON.stringify(process.env.NODE_ENV),
+          }),
+        new OptimizeCssAssetsPlugin({
+            assetNameRegExp: /\.css$/g,
+            cssProcessor: require('cssnano'),
+            cssProcessorPluginOptions: {
+              preset: ['default'],
+            },
+            canPrint: true,
+          }),
     ]
 }
